+++ src/functions/QPexFun.cc
@@ -10,7 +10,7 @@ namespace pexm {
     {}
 
     double QPexFun::scalarEval (std::vector<double const *> const &args,
-                               std::vector<unsigned int> const &dims) const
+                               std::vector<unsigned long> const &dims) const
     {
       double const * x = args[0];  // x in (0,1).
       double const * lam = args[1];
